requirements.txt
pandas
requests
beautifulsoup4
selenium
webdriver-manager
faiss-cpu
sentence-transformers
numpy
accelerate